backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlmodel==0.0.8
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pydantic==1.10.12
